LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_connection_reuse-expected.txt

CONSOLE MESSAGE: line 41: TypeError: undefined is not an object (evaluating 'entry.fetchStart')
Description

This test validates that connectStart and connectEnd are the same when a connection is reused (e.g. when a persistent connection is used).


Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'entry.fetchStart')

PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
FAIL There should be 2 PerformanceEntries assert_equals: There should be 2 PerformanceEntries expected 2 but got 0
